requirements.txt
annoy
chainlit==0.7.0
sentence-transformers==2.2.2
requests==2.31.0
python-dotenv==1.0.0
spacy==3.5.3
python-docx==0.8.11
transformers==4.30.0
torch==2.0.1